Update library to v2.0.3: Ready for Coinselector Algorithm Integration and Enhanced Documentation
This version updates the library to facilitate its use with external coinselector algorithms. The major change involves the library now returning the script satisfaction itself, instead of its size. This is particularly useful in scenarios where actual signatures are not available, as it allows the use of 72-byte zero-padded signatures (marked as 'DANGEROUSLY_USE_FAKE_SIGNATURES'). This feature should be utilized primarily for tx size estimation purposes and testing, and it's essential to be cautious with its application in real transactions.

Key Changes:
- Adapted `signatures` parameter to support 'DANGEROUSLY_USE_FAKE_SIGNATURES', enabling the generation of script satisfactions with 72-byte zero-padded signatures.
- Enhanced documentation to highlight the critical role of the `signersPubKeys` parameter in miniscript-based descriptors, especially for determining spending paths in previous outputs with various spending options.
- Included explicit warnings and guidance on the use of 'DANGEROUSLY_USE_FAKE_SIGNATURES' to prevent misuse in actual transaction scenarios.
- Comprehensive documentation updates for improved clarity and usability, including modifications to README.md and in-code comments.

These updates are designed to make the library more compatible and ready for use with coinselector algorithms, providing a more adaptable and efficient tool for Bitcoin transaction preparation and analysis.

Version: 2.0.3
Author: Jose-Luis Landabaso